+++ src/lasers/__init__.py
@@ -1,5 +1,3 @@
-from .sslaser import SolidStateLaser
-from .ndyag import NdYAGLaser
 from .cwlaser import CWLaser
 
-__all__ = ['SolidStateLaser', 'NdYAGLaser', 'CWLaser']
+__all__ = ['CWLaser']